Moved Requirement to module
moved to module and changed references
