+++ src/scripts/achievements/MoneyRequirement.ts
@@ -1,4 +1,4 @@
-///<reference path="Requirement.ts"/>
+/// <reference path="../../declarations/achievements/Requirement.d.ts" />
 
 class MoneyRequirement extends Requirement {
     constructor( requiredValue: number, type: GameConstants.AchievementOption = GameConstants.AchievementOption.more) {
