+++ src/scripts/achievements/ObtainedPokemonRequirement.ts
@@ -1,4 +1,4 @@
-///<reference path="Requirement.ts"/>
+/// <reference path="../../declarations/achievements/Requirement.d.ts" />
 
 class ObtainedPokemonRequirement extends Requirement {
     public pokemonID: number;
