+++ src/scripts/achievements/QuestLineRequirement.ts
@@ -1,4 +1,4 @@
-///<reference path="Requirement.ts"/>
+/// <reference path="../../declarations/achievements/Requirement.d.ts" />
 
 class QuestLineRequirement extends Requirement {
     private questLineName: string;
