+++ src/scripts/achievements/UndergroundItemsFoundRequirement.ts
@@ -1,4 +1,4 @@
-///<reference path="Requirement.ts"/>
+/// <reference path="../../declarations/achievements/Requirement.d.ts" />
 
 class UndergroundItemsFoundRequirement extends Requirement {
     constructor( value: number, type: GameConstants.AchievementOption = GameConstants.AchievementOption.more) {
