+++ src/scripts/achievements/WeatherRequirement.ts
@@ -1,4 +1,4 @@
-///<reference path="Requirement.ts"/>
+/// <reference path="../../declarations/achievements/Requirement.d.ts" />
 
 class WeatherRequirement extends Requirement {
